requirements.txt
flask
flask-sqlalchemy
requests
google-genai
psutil
gunicorn
flask-limiter
python-dotenv
tenacity
psycopg2-binary
midtransclient